refactor: [FE] useResultSketchbook 리팩터링
- Intent
결과스케치북을 리팩터링하면서 useResultSketchbook 훅을 지워서 결과 시작시 나오는 안내메시지가 넘어가지 않는 문제를 수정했어요.
훅 내의 페이지 이동과 스케치북 이동에 관련된 함수들을 각각 사용되는 컴포넌트로 이동했어요.
